+++ src/currency/ConvertCurrency.java
@@ -8,13 +8,13 @@ import java.io.IOException;
 
 public class ConvertCurrency {
 
-    public float getRatio(String cssQuery){
-        float ratio = 0f;
+    public double getRatio(String cssQuery){
+        double ratio = 0f;
         try {
             Document doc = Jsoup.connect("http://www.xe.com/?c=TRY").userAgent("mozilla/17.0").get();
             Elements temp = doc.select(cssQuery);
             String ratioStr = temp.text();
-            ratio = Float.parseFloat(ratioStr);
+            ratio = Double.parseDouble(ratioStr);
 
         }catch (Exception e){
             //e.printStackTrace();
